requirements.txt
google-adk
Flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-storage
google-genai
requests
streamlit
python-dotenv